refactor: split CLI into src/cli/ module and unify AgentRuntimeConfig
- Extract main.rs CLI code into src/cli/ with per-subcommand modules
  (setup, start, serve, agent, send, history, status, channel)
- Add hidden 'bridge' subcommand with --agent-id/--server-url args
  replacing env-var injection for MCP bridge invocation
- Update all driver files to pass bridge args as CLI args instead of
  env vars (claude, claude_raw, codex, codex_raw, kimi, kimi_raw,
  opencode, opencode_raw)
- Merge AcpRuntimeConfig + NativeRuntimeConfig into AgentRuntimeConfig
  using Option<String> for binary_path and acp_adaptor (None = use
  PATH / not applicable)
- All 121 lib tests passing
