+++ src/agent/drivers/codex.rs
@@ -54,8 +54,7 @@ impl AcpRuntime for CodexAcpRuntime {
             "mcpServers": [{
                 "name": "chat",
                 "command": ctx.bridge_binary,
-                "args": ["bridge", "--agent-id", ctx.agent_id, "--server-url", ctx.server_url],
-                "env": []
+                "args": ["bridge", "--agent-id", ctx.agent_id, "--server-url", ctx.server_url]
             }]
         })
     }
