+++ src/agent/drivers/codex_raw.rs
@@ -10,13 +10,11 @@ pub struct CodexRawDriver;
 
 fn build_codex_args(ctx: &SpawnContext) -> anyhow::Result<Vec<String>> {
     let bridge_binary_json = serde_json::to_string(&ctx.bridge_binary)?;
-    let bridge_args_json = serde_json::to_string(&vec![
+    let bridge_args_json = serde_json::to_string(&serde_json::json!([
         "bridge",
-        "--agent-id",
-        &ctx.agent_id,
-        "--server-url",
-        &ctx.server_url,
-    ])?;
+        "--agent-id", ctx.agent_id,
+        "--server-url", ctx.server_url,
+    ]))?;
 
     let mut args = vec!["exec".to_string()];
