+++ src/agent/drivers/kimi.rs
@@ -46,8 +46,7 @@ impl AcpRuntime for KimiAcpRuntime {
                 {
                     "name": "chat",
                     "command": ctx.bridge_binary,
-                    "args": ["bridge", "--agent-id", ctx.agent_id, "--server-url", ctx.server_url],
-                    "env": []
+                    "args": ["bridge", "--agent-id", ctx.agent_id, "--server-url", ctx.server_url]
                 }
             ]
         })
